Update Pulse to be compatible with Looker API 4.0
The pulse command was using outdated 3.0 API calls to
Looker and the try_connection sdk command is broken in the
Looker 4.0 sdk so we need to catch the exception.

This also updates the queries to run with the latest Looker
back end model and using the system_activity explores instead
of the i__looker explore.