Update cli.py
Problem Statement
Henry has been having two large issues making it difficult to use:

The Pulse command is failing because of an issue in the Looker SDK with the try_connection command.
The queries were running against the i__looker explore which seems to be getting deprecated or seems to be inaccessible to users who aren't admins.
These two problems impact Henry's ability to run and make it difficult to use. We need to fix these so that people can continue to use Henry to administer their Looker instances.

Proposed solution
Pulse command not working:

Given that the reason why the pulse command isn't working is because of the deprecation of the i__looker explore and the fact that the sdk try_connection is broken, we need to catch the exception generated by the sdk and update the queries to use the system_activity explore instead of the i__looker explore.
Analyze queries not working:

i__looker is an explore that is either deprecated or limited to admins, given that generally not all users are admins we should update to use the system_activity explore and update the fields that are used so that we can leverage this explore. An added benefit is that the system_activity explore (on a sample of 1 instance) runs much faster than i__looker.